+++ src/index.js
@@ -2,6 +2,6 @@ import './js/scroll-toggle';
 import './js/modal';
 import './js/slider';
 import './js/mobile-menu';
-import './js/header-scroll';
+import './js/page-scroll';
 import './js/modal-scroll-lock';
 import './js/snail';
