+++ src/partials/hero.html
@@ -1,4 +1,4 @@
-<section class="hero">
+<section class="hero" id="hero">
   <div class="hero__container">
     <!-- title -->
     <h1 class="hero__title">
@@ -125,4 +125,12 @@
       />
     </picture>
   </div>
+
+  <!-- backtop -->
+
+  <a href="./index.html#hero" class="backtop" data-backtop>
+    <svg class="backtop__icon">
+      <use href="./images/icons.svg#icon-arrow-short"></use>
+    </svg>
+  </a>
 </section>
