+++ src/sass/components/_backtop.scss
@@ -1,0 +1,33 @@
+.backtop {
+  position: fixed;
+  bottom: 10px;
+  right: 10px;
+  z-index: 11;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 50px;
+  height: 50px;
+
+  color: white;
+  background-color: $colorGreen;
+
+  border-radius: 50%;
+  opacity: 0.3;
+
+  @include transition(opacity);
+
+  &:hover,
+  &:focus-visible {
+    opacity: 1;
+  }
+
+  &__icon {
+    width: 50%;
+    height: 50%;
+    fill: currentColor;
+
+    transform: rotate(-90deg);
+  }
+}
