+++ src/sass/index.scss
@@ -21,6 +21,7 @@
 @import './components/mobile-menu';
 @import './components/scroll-off';
 @import './components/animation';
+@import './components/backtop';
 //
 @import './components/products';
 @import './components/about';
